[Feat] 공간 필터링 api 요청 코드 커스텀 훅으로 분리
